fai-rag-app/fai-backend/pyproject.toml
[tool.poetry]
name = "fai-backend"
version = "0.1.0"
description = ""
authors = ["Nikolas Ramstedt <nikolas.ramstedt@helsingborg.se>", "Ehsan Zilaei <ehsan.zilaei@helsingborg.se>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = { extras = ["all"], version = "^0.108.0" }
pydantic = { extras = ["email"], version = "^2.5.3" }
uvicorn = { extras = ["standard"], version = "^0.25.0" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
beanie = "^1.24.0"
mongomock = "^4.1.2"
mongomock-motor = "^0.0.26"
watchfiles = "^0.21.0"
fastapi-jwt = "^0.2.0"
rich = "^13.7.0"
requests = "^2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
